+++ src/AVLTree.ts
@@ -32,6 +32,10 @@ export class AVLTree<K extends Ord, V> implements IAVLTree<K, V> {
         return this.root === null
     }
 
+    public size(): number {
+        return this.root?.size ?? 0
+    }
+
     public minValue(): V | null {
         return minNode(this.root)?.value ?? null
     }