+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - method `IAVLTree.isEmpty()`
+- method `IAVLTree.size()`
 
 ## [0.1.1] (2021-04-25)